refactor: начать рефакторинг алгоритма переворота// Start refactoring the reversal algorithm
