+++ src/components/app/app.tsx
@@ -4,7 +4,7 @@ import { FibonacciPage } from "../fibonacci-page/fibonacci-page";
 import { ListPage } from "../list-page/list-page";
 import { MainPage } from "../main-page/main-page";
 import { QueuePage } from "../queue-page/queue-page";
-import { StringComponent } from "../string/string";
+import { StringComponent } from "../../pages/string/string";
 import { SortingPage } from "../sorting-page/sorting-page";
 import { StackPage } from "../stack-page/stack-page";
