+++ src/utils/reverse.ts
@@ -1,31 +1,44 @@
 import {ElementStates} from "../types/element-states";
-import {TItem} from "../components/string/string";
+import {TItem} from "../pages/string/string";
 import {IQueue} from "./queue";
 
 export type TCheck = Array<TItem>
 
-export function reverseWithHistory(letters: TCheck, steps: IQueue<TCheck>): any {
+export function reverseWithHistory(letters: TCheck, steps?: IQueue<TCheck>): any {
 
   let stepTemp = letters.slice();
-  let history = steps
+  let deepCopiedArray = JSON.parse(JSON.stringify(letters));
+/*  let history = steps*/
+  let history2:Array<TCheck> = []
   let startIndex = 0
   let endIndex = stepTemp.length - 1
+
+
+
+/*  history.enqueue(setChanging(stepTemp[startIndex], stepTemp[endIndex]))*/
+  history2.push(setChanging(stepTemp[startIndex], stepTemp[endIndex]))
 
 
   while (startIndex <= endIndex) {
     let start = stepTemp[startIndex]
     let end = stepTemp[endIndex]
-    history.enqueue(setChanging(start, end))
-    history.enqueue(setModified(start, end))
-    endIndex--
-    startIndex++
+
+    const nextEndIndex = endIndex - 1
+    const nextStartIndex = startIndex + 1
+    const modifiedItems = setModified(start, end)
+    const changingItems = (nextStartIndex <= nextEndIndex) ? setChanging( stepTemp[nextStartIndex], stepTemp[nextEndIndex]) : []
+ /*   history.enqueue([...modifiedItems, ...changingItems])*/
+    history2.push([...modifiedItems, ...changingItems])
+
+    endIndex = nextEndIndex
+    startIndex = nextStartIndex
   }
 
-  return history
+  return history2
 }
 
 
-function setChanging(a: any, b: any):TCheck {
+function setChanging(a: any, b: any): TCheck {
   if (a === b) {
     return [{state: ElementStates.Changing, number: a.number, value: a.value,}];
   } else {
@@ -37,7 +50,7 @@ function setChanging(a: any, b: any):TCheck {
   }
 }
 
-function setModified(a: any, b: any):TCheck {
+function setModified(a: any, b: any): TCheck {
   if (a === b) {
     return [{
       state: ElementStates.Modified,
@@ -53,9 +66,9 @@ function setModified(a: any, b: any):TCheck {
 
     }, {
 
-        state: ElementStates.Modified,
-        value: a.value,
-        number: b.number
+      state: ElementStates.Modified,
+      value: a.value,
+      number: b.number
 
     }]
   }
